+++ src/client.ts
@@ -14,20 +14,10 @@ export class Client extends Agent<IpcRenderer> {
 
   /**
    * Sends a message to the main process.
-   * @param requestChannel The channel to use for sending the request
+   * @param channel The channel to use for sending the request
    * @param data The request data
    */
-  protected send (requestChannel: string, ...data: any[]): void {
-    this.ipcService.send(requestChannel, ...data)
-  }
-
-  /**
-   * Responds to a given event from the main process.
-   * @param event The event to respond to
-   * @param responseChannel The channel to use for sending the response
-   * @param data The response data
-   */
-  protected respond (event: IpcRendererEvent, responseChannel: string, ...data: any[]): void {
-    this.ipcService.sendTo(event.senderId, responseChannel, ...data)
+  protected send (channel: string, ...data: any[]): void {
+    this.ipcService.send(channel, ...data)
   }
 }