+++ src/server.ts
@@ -15,20 +15,10 @@ export class Server extends Agent<IpcMain> {
 
   /**
    * Sends a message to the renderer process.
-   * @param requestChannel The channel to use for sending the request
+   * @param channel The channel to use for sending the request
    * @param data The request data
    */
-  protected send (requestChannel: string, ...data: any[]): void {
-    this.webContents.send(requestChannel, ...data)
-  }
-
-  /**
-   * Responds to a given event from the renderer process.
-   * @param event The event to respond to
-   * @param responseChannel The channel to use for sending the response
-   * @param data The response data
-   */
-  protected respond (event: IpcMainEvent, responseChannel: string, ...data: any[]): void {
-    event.sender.send(responseChannel, ...data)
+  protected send (channel: string, ...data: any[]): void {
+    this.webContents.send(channel, ...data)
   }
 }